chore: rename atom-binding-setup.ts to electron-binding-setup.ts
