+++ filenames.gni
@@ -54,10 +54,10 @@ filenames = {
     "lib/common/api/module-list.js",
     "lib/common/api/native-image.js",
     "lib/common/api/shell.js",
-    "lib/common/atom-binding-setup.ts",
     "lib/common/buffer-utils.js",
     "lib/common/clipboard-utils.js",
     "lib/common/crash-reporter.js",
+    "lib/common/electron-binding-setup.ts",
     "lib/common/error-utils.js",
     "lib/common/init.ts",
     "lib/common/parse-features-string.js",
