+++ lib/common/electron-binding-setup.ts
