+++ lib/content_script/init.js
@@ -4,7 +4,7 @@
 
 const { EventEmitter } = require('events')
 
-process.electronBinding = require('@electron/internal/common/atom-binding-setup').electronBindingSetup(nodeProcess._linkedBinding, 'renderer')
+process.electronBinding = require('@electron/internal/common/electron-binding-setup').electronBindingSetup(nodeProcess._linkedBinding, 'renderer')
 
 const v8Util = process.electronBinding('v8_util')
