+++ lib/sandboxed_renderer/init.js
@@ -5,7 +5,7 @@
 const events = require('events')
 const { EventEmitter } = events
 
-process.electronBinding = require('@electron/internal/common/atom-binding-setup').electronBindingSetup(binding.get, 'renderer')
+process.electronBinding = require('@electron/internal/common/electron-binding-setup').electronBindingSetup(binding.get, 'renderer')
 
 const v8Util = process.electronBinding('v8_util')
 // Expose browserify Buffer as a hidden value. This is used by C++ code to
